requirements.txt
sklearn
facenet-pytorch
wandb